feature - refactor after merge code
