+++ src/main/java/com/group7/fruitswebsite/dto/DhUserDto.java
@@ -15,14 +15,16 @@ import lombok.Setter;
 
 @Getter
 @Setter
-@JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
+@JsonIgnoreProperties(ignoreUnknown = true)
 public class DhUserDto extends BaseDto{
 	private String email;
 	private String password;
 	private String username;
 	private String avatar;
 	private String name;
-	private Set<DhRole> dhRoles = new HashSet<>();	
+	private String phone;
+	private String address;
+	private Set<DhRole> dhRoles = new HashSet<>();
 	private Set<DhProduct> dhProducts = new HashSet<>();
 	private List<DhOrder> orders = new ArrayList<>();
 }
